flea

+++ setup/fedora/config.sh
@@ -7,29 +7,44 @@ echo -e "\n\n ======= config.sh is starting ======= \n\n"
 
 # ---- dotfiles below ---------------------------
 
-dotfiles_to_symlink=($(find "$HOME/pers/dotfiles" -maxdepth 1 -mindepth 1))
-for i in "${dotfiles_to_symlink[@]}"; do
-  base_item_name=$(basename "$i")
+read -rp "Do you want to link dotfiles? (y/n): " dotfiles_res
+if [[ "$dotfiles_res" =~ ^[Yy]$ ]]; then
+  dotfiles_to_symlink=($(find "$HOME/pers/dotfiles" -maxdepth 1 -mindepth 1))
+  for i in "${dotfiles_to_symlink[@]}"; do
+    base_item_name=$(basename "$i")
 
-  echo "Linking $i to $HOME/$base_item_name"
-  rm -rf "$HOME/$base_item_name"
-  ln -sf "$i" "$HOME/$base_item_name"
-done
+    echo "Linking $i to $HOME/$base_item_name"
+    rm -rf "$HOME/$base_item_name"
+    ln -sf "$i" "$HOME/$base_item_name"
+  done
+
+  # omz config below
+
+  cd "$HOME/pers/dotfiles/.oh-my-zsh"
+  git submodule update --init --recursive
+  cd -
+fi
 
 # ---- secrets below ---------------------------
 
-find "$HOME/pers/secrets" -type f -exec ansible-vault decrypt --vault-password-file "$HOME/pers/password" -- {} \;
+read -rp "Do you want to link secrets? (y/n): " secrets_res
+if [[ "$secrets_res" =~ ^[Yy]$ ]]; then
+  find "$HOME/pers/secrets" -type f -exec ansible-vault decrypt --vault-password-file "$HOME/pers/password" -- {} \;
 
-mv "$HOME/.ssh" "$HOME/.ssh.backup"
-ln -sf "$HOME/pers/secrets/ssh" "$HOME/.ssh"
+  mv "$HOME/.ssh" "$HOME/.ssh.backup"
+  ln -sf "$HOME/pers/secrets/ssh" "$HOME/.ssh"
 
-mv "$HOME/.zsh_history" "$HOME/.zsh_history.backup"
-ln -sf "$HOME/pers/secrets/.zsh_history" "$HOME/.zsh_history"
+  mv "$HOME/.zsh_history" "$HOME/.zsh_history.backup"
+  ln -sf "$HOME/pers/secrets/.zsh_history" "$HOME/.zsh_history"
+fi
 
 # ---- networking below ---------------------------
 
-read -rp "Enter pretty hostname: " pretty_hostname
-sudo hostnamectl set-hostname --pretty "$pretty_hostname"
+read -rp "Do you want to tweak networking settings? (y/n): " networking_res
+if [[ "$networking_res" =~ ^[Yy]$ ]]; then
+  read -rp "Enter pretty hostname: " pretty_hostname
+  sudo hostnamectl set-hostname --pretty "$pretty_hostname"
 
-read -rp "Enter static hostname: " static_hostname
-sudo hostnamectl set-hostname --static "$static_hostname"
+  read -rp "Enter static hostname: " static_hostname
+  sudo hostnamectl set-hostname --static "$static_hostname"
+fi
